(Android) Make test-app work without new RN arch
